+++ dynamo3/__init__.py
@@ -9,6 +9,7 @@ from .exception import (CheckFailed, ConditionalCheckFailedException,
                         ProvisionedThroughputExceededException, DynamoDBError)
 from .fields import (Throughput, Table, DynamoKey, LocalIndex, GlobalIndex,
                      IndexUpdate)
+from .result import Limit
 from .types import Dynamizer, Binary, TYPES, TYPES_REV, is_null
 
-__version__ = '0.4.5'
+__version__ = '0.4.6'